Create Utility class.
