+++ genetic_algorithm/five_parameters/Utility.cpp
@@ -1,0 +1,44 @@
+#include "Utility.h"
+
+unsigned long long GetTimeStamp()
+{
+	namespace pt = boost::posix_time;
+	namespace gg = boost::gregorian;
+
+	typedef boost::date_time::c_local_adjustor<pt::ptime> local_adj;
+
+	auto epoch = local_adj::utc_to_local(pt::ptime(gg::date(1970, 1, 1)));
+	auto diff = pt::microsec_clock::local_time() - epoch;
+
+	return diff.total_microseconds();
+}
+
+std::string GetTimeISOString()
+{
+	boost::posix_time::ptime today = boost::posix_time::second_clock::local_time();
+
+	return boost::posix_time::to_iso_string(today);
+}
+
+void Random(int min, int max, int result[][PARAMETER_NUM], int num)
+{
+	boost::random::mt19937 gen(GetTimeStamp());
+	boost::random::uniform_int_distribution<> dist(min, max);
+
+	boost::this_thread::sleep(boost::posix_time::microseconds(1000));
+
+	for(int i=0; i<RANDOM_MAX; i++)
+	{
+		result[i][num] = dist(gen);
+	}
+}
+
+int Random(int min, int max)
+{
+	boost::random::mt19937 gen(GetTimeStamp());
+	boost::random::uniform_int_distribution<> dist(min, max);
+
+	boost::this_thread::sleep(boost::posix_time::microseconds(1000));
+
+	return dist(gen);
+}
